backend/requirements.txt
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
email-validator==2.1.0.post1
gunicorn==21.2.0
ipython==8.21.0
mysqlclient==2.2.1
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
virtualenv==20.25.0
Werkzeug==3.0.1
WTForms==3.1.2
